pyproject.toml
[project]
name = "credit-ocr-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.7.0",
    "camelot-py[cv]>=1.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "huggingface-hub[hf-xet]>=0.31.2",
    "ipywidgets>=8.1.7",
    "ollama>=0.4.8",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
]
